Added user auth, user model and auth handlers; modified record table and controller to include user id created_by

+++ controller.go
@@ -14,10 +14,11 @@ type dataRecord struct {
 }
 
 // func to get all records
-func getRecords() ([]dataRecord, error) {
+func getRecords(userId string) ([]dataRecord, error) {
 	var records []dataRecord
+	var recUserId string
 
-	rows, err := db.Query("SELECT * FROM records")
+	rows, err := db.Query("SELECT * FROM records WHERE created_by = $1", userId)
 	if err != nil {
 		return nil, fmt.Errorf("error getting records: %v", err)
 	}
@@ -27,7 +28,7 @@ func getRecords() ([]dataRecord, error) {
 	for rows.Next() {
 		var rec dataRecord
 
-		if err := rows.Scan(&rec.Id, &rec.Title, &rec.Comment, &rec.LastDate); err != nil {
+		if err := rows.Scan(&rec.Id, &rec.Title, &rec.Comment, &rec.LastDate, &recUserId); err != nil {
 			return nil, fmt.Errorf("error reading records: %v", err)
 		}
 
@@ -37,11 +38,12 @@ func getRecords() ([]dataRecord, error) {
 	return records, nil
 }
 
-func getRecordById(recordId string) (dataRecord, error) {
+func getRecordById(recordId string, userId string) (dataRecord, error) {
 	var rec dataRecord
+	var recUserId string
 
-	row := db.QueryRow("SELECT * FROM records WHERE id = $1", recordId)
-	if err := row.Scan(&rec.Id, &rec.Title, &rec.Comment, &rec.LastDate); err != nil {
+	row := db.QueryRow("SELECT * FROM records WHERE id = $1 AND created_by = $2", recordId, userId)
+	if err := row.Scan(&rec.Id, &rec.Title, &rec.Comment, &rec.LastDate, &recUserId); err != nil {
 		if err == sql.ErrNoRows {
 			return rec, fmt.Errorf("recordById %s: no such record", recordId)
 		}
@@ -50,9 +52,9 @@ func getRecordById(recordId string) (dataRecord, error) {
 	return rec, nil
 }
 
-func createRecord(record dataRecord) (string, error) {
+func createRecord(record dataRecord, userId string) (string, error) {
 
-	result, err := db.Exec("INSERT INTO records (title, comment, last_date) VALUES ($1, $2, to_timestamp($3, 'YYYY-MM-DD'))", record.Title, record.Comment, record.LastDate)
+	result, err := db.Exec("INSERT INTO records (title, comment, last_date, created_by) VALUES ($1, $2, to_timestamp($3, 'YYYY-MM-DD'), $4)", record.Title, record.Comment, record.LastDate, userId)
 	if err != nil {
 		return "0", fmt.Errorf("addRecord: %v", err)
 	}
@@ -63,9 +65,9 @@ func createRecord(record dataRecord) (string, error) {
 	return string(id), nil
 }
 
-func updateRecord(record dataRecord) (string, error) {
+func updateRecord(record dataRecord, userId string) (string, error) {
 
-	result, err := db.Exec("UPDATE records SET title = $1, comment = $2, last_date = to_timestamp($3, 'YYYY-MM-DD') WHERE id = $4", record.Title, record.Comment, record.LastDate, record.Id)
+	result, err := db.Exec("UPDATE records SET title = $1, comment = $2, last_date = to_timestamp($3, 'YYYY-MM-DD') WHERE id = $4 AND created_by = $5", record.Title, record.Comment, record.LastDate, record.Id, userId)
 	if err != nil {
 		return "0", fmt.Errorf("addRecord: %v", err)
 	}
@@ -76,8 +78,8 @@ func updateRecord(record dataRecord) (string, error) {
 	return string(id), nil
 }
 
-func deleteRecord(recordId string) (string, error) {
-	_, err := db.Exec("DELETE FROM records WHERE id = $1", recordId)
+func deleteRecord(recordId string, userId string) (string, error) {
+	_, err := db.Exec("DELETE FROM records WHERE id = $1 AND created_by = $2", recordId, userId)
 
 	if err != nil {
 		return "0", fmt.Errorf("addRecord: %v", err)
